The options move out of feed.go, and a waiver finds its subject again
Two things the rebase onto main surfaced.

`feed.go` crossed the 500-line cap once the lead lane and the board's
counting reader both landed in it. The `With…` options move to
`feedoptions.go` — they are one concept, they are what a reader adding a
lane comes here for, and neither file is near the cap now.

`TestEveryReaderOfTheActivityTableExcludesRestrictedRows` is RED ON MAIN,
and this fixes it. `ListAuditLog`'s query moved to
`auditaudienceboundary.go`; the gate attributes a reader to the file
holding its SQL, so the waiver's key still named `auditlog.go` and matched
nothing. The gate reported both halves — an unwaived reader and a waiver
matching no subject — which is what a census that cannot fail short looks
like when it works.

The waiver's reasoning is unchanged, and it still guards the real
subject: pointing the key at a file that does not exist fails the gate.
